Reduce duplication and fix incorrect length/precision
